Ready for first deployment

+++ collabl/groups/views_htmx.py
@@ -1,17 +1,23 @@
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseForbidden, HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404
 from django.urls import reverse_lazy
+from django.views.decorators.http import require_http_methods
 
 import groups.constants as c
 from collabl.settings import SITE_PROTOCOL, SITE_DOMAIN
 from groups.forms import GroupForm, GroupImageForm, GroupAnnouncementForm
 from groups.models import Group, Membership, GroupAnnouncement
-from groups.utils import get_membership_level, get_filtered_collaborations
+from groups.utils import (
+    get_membership_level,
+    get_filtered_collaborations,
+    user_is_admin,
+)
 from groups.views import get_membership_count
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
 def group_update_view(request, slug):
     """
     HTMX VIEW - Allows group updates with no reload
@@ -20,15 +26,15 @@ def group_update_view(request, slug):
     (with error messages, if appropriate)
     """
 
+    # Get Data
     group = get_object_or_404(Group, slug=slug)
-
     form = GroupForm(request.POST or None, instance=group)
 
-    if request.user.is_authenticated:
-        membership_level = get_membership_level(request.user, group)
-    else:
-        membership_level = None
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
 
+    # If POST, process the update
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -37,11 +43,12 @@ def group_update_view(request, slug):
             "app/group/partials/header/main.html",
             {
                 "group": group,
-                "membership_level": membership_level,
+                "membership_level": get_membership_level(request.user, group),
                 "membership_count": get_membership_count(group),
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Modal
     return render(
         request,
         "app/group/partials/modals/group_update.html",
@@ -53,6 +60,32 @@ def group_update_view(request, slug):
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
+def group_delete_view(request, slug):
+    """
+    HTMX VIEW - Allows group deletion
+    On GET, sends back confirmation modal
+    ON POST, deletes the group, and redirects user to the group list page.
+    """
+
+    # Get Data
+    group = get_object_or_404(Group, slug=slug)
+
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
+
+    # If POST, delete the group
+    if request.method == "POST":
+        group.delete()
+        return HttpResponseRedirect(reverse_lazy("user-group-list"))
+
+    # If GET, (or invalid data is posted) send back the Modal
+    return render(request,"app/group/partials/modals/group_delete.html",{"group": group,},)
+
+
+@login_required()
+@require_http_methods(["GET", "POST"])
 def group_image_view(request, slug):
     """
     HTMX VIEW - Allows group image updates with no reload
@@ -61,15 +94,15 @@ def group_image_view(request, slug):
     (with error messages, if appropriate)
     """
 
+    # Get Data
     group = get_object_or_404(Group, slug=slug)
-
     form = GroupImageForm(request.POST or None, request.FILES or None, instance=group)
 
-    if request.user.is_authenticated:
-        membership_level = get_membership_level(request.user, group)
-    else:
-        membership_level = None
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
 
+    # If POST, process the update
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -78,11 +111,12 @@ def group_image_view(request, slug):
             "app/group/partials/header/main.html",
             {
                 "group": group,
-                "membership_level": membership_level,
+                "membership_level": get_membership_level(request.user, group),
                 "membership_count": get_membership_count(group),
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Modal
     return render(
         request,
         "app/group/partials/modals/group_image.html",
@@ -94,39 +128,50 @@ def group_image_view(request, slug):
 
 
 @login_required()
+@require_http_methods(["GET",])
 def group_membership_view(request, slug):
     """
     HTMX VIEW - Populates list of memberships of the specified type - set by select object on front end
+
+    if the filter is not set, we send back only pending membership requests
     """
 
-    # if the filter is not set, we send back only pending membership requests
+    # Get Data & Validate
+    group = get_object_or_404(Group, slug=slug)
     membership_filter = request.GET.get(
         "membership_filter", c.MEMBERSHIP_STATUS_PENDING
     )
-    group = get_object_or_404(Group, slug=slug)
+
+    # If the filter isn't out list of membership statuses, we dont want it thrown at the database.
+    # Instead, we return nothing. (This also acts as a 'HIDE' function on the FE)
+    if membership_filter not in c.MEMBERSHIP_FILTERS:
+        return HttpResponse("")
+
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
 
     # Clear the session, if it is being used
     if request.session.get("selected_memberships", None):
         del request.session["selected_memberships"]
 
-    if membership_filter in c.MEMBERSHIP_FILTERS:
-        return render(
-            request,
-            "app/group/partials/memberships/list.html",
-            {
-                "membership_list": Membership.objects.filter(
-                    group__slug=slug, status=membership_filter
-                ),
-                "membership_filter": membership_filter,
-                "group": group,
-                "membership_count": get_membership_count(group),
-            },
-        )
-    else:
-        return HttpResponse()
+    # filter the queryset and send back the rendered template
+    return render(
+        request,
+        "app/group/partials/memberships/list.html",
+        {
+            "membership_list": Membership.objects.filter(
+                group__slug=slug, status=membership_filter
+            ),
+            "membership_filter": membership_filter,
+            "group": group,
+            "membership_count": get_membership_count(group),
+        },
+    )
 
 
 @login_required()
+@require_http_methods(["POST",])
 def group_membership_selector_view(request, slug, pk, membership_filter):
     """
     HTMX VIEW - Allows admins to select memberships in order to process in bulk
@@ -141,6 +186,10 @@ def group_membership_selector_view(request, slug, pk, membership_filter):
     # Grab the current list, or create one - an empty list
     selected_memberships = request.session.get("selected_memberships", [])
     group = get_object_or_404(Group, slug=slug)
+
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
 
     # CASE 1: Removing an agency - Update the shortlist (in session), render the response
     if pk in selected_memberships:
@@ -176,6 +225,7 @@ def group_membership_selector_view(request, slug, pk, membership_filter):
 
 
 @login_required()
+@require_http_methods(["POST",])
 def group_membership_handler_view(request, slug, action, membership_filter):
     """
     HTMX VIEW - Allows admins process memberships stored in session
@@ -187,12 +237,18 @@ def group_membership_handler_view(request, slug, action, membership_filter):
     count or check/uncheck boxes)
     """
 
+    # Get Data
+    group = get_object_or_404(Group, slug=slug)
+
     # Grab the current list, or return None is there isn't one
     if not (selected_memberships := request.session.get("selected_memberships", None)):
         return None
 
-    group = get_object_or_404(Group, slug=slug)
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
 
+    # Process the request (depending on the action specified)
     if action == c.MEMBERSHIP_ACTION_CLEAR_SELECTION:
         # Get the ids (so that we can 'uncheck' the checkboxes on front end)
         check_box_ids = [
@@ -244,6 +300,7 @@ def group_membership_handler_view(request, slug, action, membership_filter):
     else:
         membership_list = Membership.objects.none()
 
+    # Make Response (updating state of the memberships section)
     return render(
         request,
         "app/group/partials/memberships/main.html",
@@ -257,6 +314,7 @@ def group_membership_handler_view(request, slug, action, membership_filter):
 
 
 @login_required()
+@require_http_methods(["GET",])
 def group_collaboration_list(request, slug):
     """
     HTMX VIEW - Populates the list of collaborations - either All, Planning, ongoing,
@@ -270,28 +328,33 @@ def group_collaboration_list(request, slug):
     if collaboration_list_filter == "HIDE":
         return HttpResponse()
 
-    collaborations = get_filtered_collaborations(group, collaboration_list_filter)
-
+    # Make Response
     return render(
         request,
         "app/group/partials/collaborations/list.html",
-        {"collaboration_list": collaborations, "group": group},
+        {
+            "collaboration_list": get_filtered_collaborations(
+                group, collaboration_list_filter
+            ),
+            "group": group,
+        },
     )
 
 
 @login_required()
+@require_http_methods(["GET",])
 def group_announcement_list(request, slug):
     """
     HTMX VIEW - Populates the list of announcements - either Latest, All, or None
     """
 
-    # if the filter is not set, we hide the announcements
+    # if the filter is not set (ir is set to hide), we hide the announcements
     announcement_list_filter = request.GET.get("announcement_list_filter", "HIDE")
-    group = get_object_or_404(Group, slug=slug)
-
     if announcement_list_filter == "HIDE":
-        return HttpResponse()
+        return HttpResponse("")
 
+    # Filter the announcements
+    group = get_object_or_404(Group, slug=slug)
     match announcement_list_filter:
         case c.ANNOUNCEMENTS_FILTER_LATEST:
             announcements = GroupAnnouncement.objects.filter(group=group)[:1]
@@ -300,6 +363,7 @@ def group_announcement_list(request, slug):
         case _:
             announcements = GroupAnnouncement.objects.none()
 
+    # Make Response
     return render(
         request,
         "app/group/partials/announcements/list.html",
@@ -308,6 +372,7 @@ def group_announcement_list(request, slug):
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
 def group_announcement_delete(request, slug, pk):
     """
     HTMX VIEW - Allows announcement updates with no reload
@@ -316,9 +381,15 @@ def group_announcement_delete(request, slug, pk):
     (with error messages, if appropriate)
     """
 
+    # Get Data
     group = get_object_or_404(Group, slug=slug)
     announcement = get_object_or_404(GroupAnnouncement, pk=pk)
 
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
+
+    # If POST, process the delete operation
     if request.method == "POST":
         announcement.delete()
         return render(
@@ -330,6 +401,7 @@ def group_announcement_delete(request, slug, pk):
             },
         )
 
+    # If GET, send back the 'Delete Announcement' Modal
     return render(
         request,
         "app/group/partials/announcements/list.html",
@@ -343,6 +415,7 @@ def group_announcement_delete(request, slug, pk):
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
 def group_announcement_create(request, slug):
     """
     HTMX VIEW - Allows new announcements with no reload
@@ -351,10 +424,15 @@ def group_announcement_create(request, slug):
     (with error messages, if appropriate)
     """
 
+    # Get Data
     group = get_object_or_404(Group, slug=slug)
-
     form = GroupAnnouncementForm(request.POST or None)
 
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
+
+    # If POST, process the operation
     if request.method == "POST" and form.is_valid():
         announcement = form.save(commit=False)
         announcement.group = group
@@ -369,6 +447,7 @@ def group_announcement_create(request, slug):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the 'Create Announcement' Modal
     return render(
         request,
         "app/group/partials/announcements/list.html",
@@ -382,6 +461,7 @@ def group_announcement_create(request, slug):
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
 def group_announcement_update(request, slug, pk):
     """
     HTMX VIEW - Allows announcement updates with no reload
@@ -390,13 +470,20 @@ def group_announcement_update(request, slug, pk):
     (with error messages, if appropriate)
     """
 
+    # Get Data
     group = get_object_or_404(Group, slug=slug)
     group_announcement = get_object_or_404(GroupAnnouncement, pk=pk)
-
     form = GroupAnnouncementForm(request.POST or None, instance=group_announcement)
 
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
+
+    # If POST, process the update (and change the user, if needed)
     if request.method == "POST" and form.is_valid():
-        form.save()
+        group_announcement = form.save(commit=False)
+        group_announcement.user = request.user
+        group_announcement.save()
         return render(
             request,
             "app/group/partials/announcements/list.html",
@@ -406,6 +493,7 @@ def group_announcement_update(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the 'Update' Modal
     return render(
         request,
         "app/group/partials/announcements/list.html",
@@ -420,6 +508,7 @@ def group_announcement_update(request, slug, pk):
 
 
 @login_required()
+@require_http_methods(["GET", "POST"])
 def group_create_view(request):
     """
     HTMX VIEW - Allows group creation
@@ -427,15 +516,24 @@ def group_create_view(request):
     on successful submission, sends back a JS/htmx redirect to the new group page.
     """
 
+    # Get data
     form = GroupForm(request.POST or None)
+
+    # If POST, create the group
     if request.method == "POST":
         if form.is_valid():
+
+            # Create the group
             group = form.save(commit=False)
             group.created_by = request.user
             group.save()
+
+            # Add the user as Admin
             Membership.objects.create(
                 user=request.user, group=group, status=c.MEMBERSHIP_STATUS_ADMIN
             )
+
+            # Get success_url - we send back a javascript redirect to take the user to this page
             success_url = (
                 SITE_PROTOCOL
                 + SITE_DOMAIN
@@ -445,6 +543,7 @@ def group_create_view(request):
                 )
             )
 
+            # Make Response
             return render(
                 request,
                 "app/snippets/js_redirect.html",
@@ -453,6 +552,7 @@ def group_create_view(request):
                 },
             )
 
+    # If GET, (or invalid data is posted) send back the 'Create Group' Modal
     return render(
         request,
         "app/home/modals/group_create.html",
